[fix] 타입 파일 별로 나눠서 types 폴더로 이동
