+++ constants/index.ts
@@ -1,5 +1,4 @@
-export type GroupColor = '#f86565' | '#f89a65' | '#f8e065' | '#658ef8' | '#9465f8';
-export type GroupType = '스터디' | '학교, 교내/외 모임' | '회사, 사내 모임' | '취미, 동호회 모임' | '친구, 사모임' | '프로젝트' | '기타' | '';
+import { GroupColor, GroupType } from '@/types/group';
 
 export const PLACEHOLDER = {
   NAME: '모임에서 사용할 이름을 입력해 주세요.',
