+++ queries/Detail/useCreateDetail.ts
@@ -1,8 +1,6 @@
-import { createDetail, DetailInfo } from '@/api/Detail';
-import { ServerResponse } from '@/types';
+import { createDetail } from '@/api/Detail';
 import { useMutation } from '@tanstack/react-query';
-import { AxiosError } from 'axios';
 
 export const useCreateDetail = () => {
-  return useMutation<ServerResponse<{ id: string }>, AxiosError, DetailInfo>(createDetail);
+  return useMutation(createDetail);
 };
