+++ queries/Detail/useUpdateDetailStatus.ts
@@ -1,8 +1,6 @@
-import { DetailInfo, DetailStatusWithEventId, updateDetailStatus } from '@/api/Detail';
-import { AxiosError } from 'axios';
-import { ServerResponse } from '@/types';
+import { updateDetailStatus } from '@/api/Detail';
 import { useMutation } from '@tanstack/react-query';
 
 export const useUpdateDetailStatus = () => {
-  return useMutation<ServerResponse<DetailInfo>, AxiosError, DetailStatusWithEventId>(updateDetailStatus);
+  return useMutation(updateDetailStatus);
 };
