+++ queries/Group/useCreateGroup.ts
@@ -1,9 +1,7 @@
 import { message } from './index';
-import { GroupId, createGroup, deleteGroup, GroupInfo, joinGroup, GroupNickname, withdrawalGroup, changeNickname, changeAdmin } from '@/api/Group';
-import { useMutation, useQuery } from '@tanstack/react-query';
-import { AxiosError } from 'axios';
-import { ServerResponse } from '@/types';
+import { createGroup } from '@/api/Group';
+import { useMutation } from '@tanstack/react-query';
 
 export const useCreateGroup = () => {
-  return useMutation<ServerResponse<GroupId>, AxiosError, GroupInfo>(createGroup, message);
+  return useMutation(createGroup, message);
 };
