+++ queries/Group/useGroupDetail.ts
@@ -1,7 +1,6 @@
-import { getGroupDetail, GroupDetail } from '@/api/Group';
-import { ServerResponse } from '@/types';
+import { getGroupDetail } from '@/api/Group';
 import { useQuery } from '@tanstack/react-query';
 
 export const useGroupDetail = (groupId: string) => {
-  return useQuery<ServerResponse<GroupDetail>>(['groupDetail', groupId], () => getGroupDetail(groupId));
+  return useQuery(['groupDetail', groupId], () => getGroupDetail(groupId));
 };
