+++ queries/Group/useParticipantList.ts
@@ -1,7 +1,6 @@
-import { getParticipantList, ParticipantList } from '@/api/Group';
-import { ServerResponse } from '@/types';
+import { getParticipantList } from '@/api/Group';
 import { useQuery } from '@tanstack/react-query';
 
 export const useParticipantList = (groupId: string) => {
-  return useQuery<ServerResponse<ParticipantList>>(['participantList', groupId], () => getParticipantList(groupId));
+  return useQuery(['participantList', groupId], () => getParticipantList(groupId));
 };
